pyproject.toml
[project]
name = "email-opening-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "msal (>=1.31.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
